refactor: remove legacy agent_tasks imports
