+++ api/src/app/routes/agents.py
@@ -4,8 +4,8 @@ from fastapi import APIRouter, HTTPException
 from postgrest.exceptions import APIError
 from pydantic import BaseModel
 
-from ..agent_tasks.infra.infra_cil_validator_agent import run as run_infra_cil_validator
 from app.agents.runtime.infra_observer_agent import run as run_orch_block_manager
+
 from ..utils.supabase_client import supabase_client as supabase
 
 router = APIRouter(prefix="/agents", tags=["agents"])
@@ -29,8 +29,6 @@ def run_agent(name: str, payload: AgentRunPayload):
     try:
         if name == "orch_block_manager":
             result = run_orch_block_manager(payload.basket_id)
-        elif name == "infra_cil_validator":
-            result = run_infra_cil_validator(payload.basket_id)
         else:
             raise HTTPException(status_code=404, detail="unknown agent")
     except APIError as err:
