+++ src/app/agents/runtime/infra_research_agent.py
@@ -5,10 +5,10 @@ from datetime import datetime, timezone
 import asyncpg
 from schemas.research import ResearchIn, ResearchOut
 from schemas.validators import validates
-from src.app.agent_tasks.layer1_infra.utils.block_policy import insert_revision, is_auto
-from src.utils.logged_agent import logged
 
+from app.agents.utils.block_policy import insert_revision, is_auto
 from app.event_bus import DATABASE_URL, publish_event
+from src.utils.logged_agent import logged
 
 from ..schemas import RefreshReport
